Separate model class test and tests for property models
